Replace two models with `YearlyStatisticsModel` - Replace `StatisticsHistoryModel` and `ProductsStatisticsModel` with `YearlyStatisticsModel`.
- Add two properties to `YearlyStatisticsSchema`, `categories` and
`monthes`, Rename `statisticType` property to `statisticsType` .

- Replace `statistics_history` file with `monthly_statistics` file
due to `StatisticsHistoryModel` that is used in `statistics_history`
has been deprecated now. and the new file `monthly_statistics`
uses the new model `YearlyStatisticsModel` to get the yearly statistics
of the store.

- Replace `ProductsStatisticsModel` and the complexited logec that
comes after it with `YearlyStatisticsModel` thats return ready object
contains the statistics of categories in `getCategoriesStatistics` file