Separate signing up logic from `user_regist` Separate signing up logic from `user_regist` function to a separated function named `users_signUpUser`
- Create `checkEmailExistance` function and add it to `user_regist` function
for checks if the given email from the user who want to sign up is exist and active email or not.

- Change return values from `user_regist` function to the client by
make the response an object with `ok` property tells if the process done successfully
or no, and `message` property to descripe the error if happend,
and `payload` propery to return data with the response.