Create users emails verification logic 1 - Create `/email-verification` endpoint for recieveing verifying requestes.
2 - `userEmailVerification` function recives verification requestes and generates
verification code using `getRandomInt` function and sends it to requester's email
using `sendVerificationEmailMail` function.

3 - `sendVerificationEmailMail` function when called to send the code to the user,
it uses `emailHtmlTemplate` function to create html template for pass it with
`sendEmail` function to user email.

4 - If sending verification code done successfully, `verifyUserEmail` will wates the user
to opens its email and brings verification code and sends it as post request to the endpoint
`/email-verification`, therefore, `verifyUserEmail` will checks if the given code equals to that
we sent or no, if equals, so user email verified.

- Suspension `/mails` endpoint and function `sendFeedback` temporarily.