Create `ErrorGenerator` class Create `ErrorGenerator` class for building a centralized place to handle preparing the errors and sending them to the client.
- Create the a function named `findErrorPlace` takes `stack` property from
`ErrorGenerator` class (whome inherits the property from `Error` class) and
extracts the name of the file that the error came from it, this function
will used only in development environment for helps the developer to know
from where the error came.

- install `cross-env` package for setting on the fly the value of
`NODE_ENV` environment variable.

- Set an listener function catch the unhandled rejection errors and shuts down
the server (those unhandled rejection error only will be the processes that
the api depends on them for work fine.) so the others processes will be catched
and handled.

- Rename `functions` folder to `utilities`.