Refactor the tests files of products routes
- Use the functions `userRequest`, `adminRequest` and `anyRequest` to send
tests requests instead of using `supertest` package and the instance of the server
from `/src/server.js` file directly to send testing requests.

- Use `closeTestingServer` function to close the connection to
the database and close the running server in all tests files that use
`mongoose` package directly to close the connection to the database
and use the instance of the server from `/src/server.js` file to close the server.