Upload docker compose files
- Remove "docker-compose*" line from `.gitignore` file for
allowing docker compose files being pushed to the repository.

- Upload four docker compose files:
1 - `docker-compose-dev.yml` for defining and running the application
or the server in development environment.

2 - `docker-compose-prod.yml` for defining and running the application
or the server in production environment.

3 - `docker-compose-test.yml` that defines and runs the application
or the server in testing environment.

4 - `docker-compose-test-replicas.yml` for defining and running three
mongo databases in local machine as a replica set for use it in
jest integration testing.

- Create `scripts` folder and create inside it the script that configures
and initializes the replica set of mongo databases and the the script
that starts or runs the containers of replica set databases.

- Create scripts/mongodb-replicas-init.sh
 configurations and the
initialization of mongo databases replica set of `docker-compose-test-replicas.yml`