Refactor certain tests files and mock some utilities
- Refactor `forgetPassword_post.test.js` file, Mock `genRandomNumber` function
that generates random verification codes and make it returns `999999` always,
Add more tests for more ensuring that the process of changing user's password
done successfully by checking that the function (`sendEmail`) which will sends
a mail to the user to tell him that his password changed successfully called,
So mock `sendEmail` function and make it returns `true` directly which means
that the email sent successfully.

- Refactor `register_user_post.test.js` file, Mock `checkEmailExistance` function
for skiping the http request that will sent for checking if the provided email
is exist email of not, Add another test case with "Should returns an error with
message 'Your email is not active'" description for test the case if a not exist
email provided.

- Refactor `orders_addNewOrder_post.test.js` file, Mock `sendEmail` function
(that will called after creating the order to sends a mail to the user) to make it
returns `true` directly without sending a real email, Create `doesEmailMessageSent`
function for reusing the part of test that checks if `sendEmail` function are called
and called with which arrguments.

- Refactor the second test case in `user_paymentMethods_get.test.js` file.

- Cancel logging the whole error object in `scripts/jest-runner-script.js`

- Change the type of `expired` property in payment method schema
to `String` inside `src/models/Users.js` file.

- Add `"testEnvironment": "node"` property to `jest.config.json` file.