Cache top products list & add caching option to search function
- Use `cacheOrQuery` function in `products_topProducts_get` route handler
to cache top products list for one hour and then revalidate the cache.

- Add caching option to `products_userSearch_get` route handler through
`cacheFor` query parameter, the query parameter should consist of a strings
of two parts separated by `-`, the first part represents the type of the time
(E.g, hours, minutes, weeks, ect...), and the second part represents the
number of the given time in the first part.