feat: create a function to update data in redis cache
Create `updateRedisCache` function to provide an easy way to update data
in redis cache.

Make `updateSetting` function takes an instant of `ClientSession` class
as option.

Use MongoDB session through `startSession` and Implement a logic in
`settings_updateSetting_post` to ensure that the database and redis cache updated
successfully in case the setting that is targeted to update stored in redis cache.