fix: prevent `createCategory` function from adding an existing category
Change `$push` operation to `$addToSet` in `createCategory` function for avoiding
duplicating existing categories, And make the function returns `null` if the incoming
category is already exist.

Add "if" statement to `products_createCategory_post` route handler to check if
the respond of `createCategory` function equals `null` for returning an error
with "The category already exist" message.