refactor: change the way of setting expiration date for search result cache
Set a fixed expiration date for products search result cache instead of using `cacheFor`
query parameter to specify the expiration date, That is in order not to give the client
side the ability to set the expiration date for products search result cache.

Use `useCache` query parameter to determine if the client wants to get the cached
search result if exist or wants to get a fresh data from the database directly.