ci: create a custom github action for collecting tests code coverage report
Create `tests-code-coverage-report` folder and create inside it `action.yml` file,
which is the file that defines the custom github action and specifies the scripts
that will run when the action called.

Create `collecting-tests-code-coverage-report-scripts.js` file in `tests-code-coverage-report`
folder, which is the file that contains the scripts that will run and handle collecting tests
code coverage report in one folder to deploy it to reports repository.

Add `coverage-badges-cli` & `@actions/core` packages as dev dependencies
to be used by the scripts in `collecting-tests-code-coverage-report-scripts.js` file.

Add `tests-code-coverage-report` folder to `.dockerignore` file.