feat: set default documents for three collections
When the server run for the first time and the database is empty :point_down::

- Make `SettingsModel` in `Settings.js` file creates a settings document with the
default settings values.

- Make `YearlyStatisticsMode` in `YearlyStatistics.js` file creates yearly statistics
document for the current year with the default values.

- Make `AdminsModel` in `Admins.js` file creates a default admin account in
admins collection.

Create `defaultAdmin.js` file and export from it the data of the default admin
account.

Refactor how `logIn_loggedAdmin_get` route handler sends the responses.