chore: create a reusable mongodb replica set containers
*Why replica set and not a single mongo container?*
Because the application uses transactions feature and this feature only
works in replica set environment.

Rename `docker-compose-test-replicas.yml` file to
`docker-compose-mongodb-replSet.yml` and set up inside it the mongodb
replica set for using it in development or testing environments.

Create `run-mongodb-replSet.js` file that contains the script that runs and
initializes the mongodb replica set, And delete `mongodb-replicas-start.sh`
script file because `run-mongodb-replSet.js` toke its job.

Rename `mongodb-replicas-init.sh` script file that is used to initializing
mongodb replica set to `initialize-mongodb-replSet.sh` to be suitable
for `docker-compose-test-replicas.yml` & `run-mongodb-replSet.js`.

Add `mongodb` folder to `.gitignore` and `.dockerignore` files because
this folder will used to store the data of the localy running database.

Delete `docker-compose-test.yml` file because it unnecessary currently.