feat: create some initialization functions should run before the server
Convert the line of code that connects to the database in `databaseConnections`
file to a function called `databaseConnections` and export it.

Convert the line of code that connects to the database in `redisClient.js`
file to a function called `redisClientConnection` and export it.

Convert the lines of code that create the default settings document if
settings collection is empty in `Settings.js` file to a function called
`initializeSettingsCollection` and export it.

Convert the lines of code that create a yearly statistics document for the
current year if yearly statistics collection is empty in `YearlyStatistics.js`
file to a function called `initializeYearlyStatisticsCollection` and export it.

Create `initializations` function in `configuration/index.js` file and collect
inside it the processes that need to be done before starting the server
which them: `databaseConnections`, `redisClientConnection`,
`initializeSettingsCollection` and `initializeYearlyStatisticsCollection`.